Fix NPE when listing directory; patch by yukim reviewed by jbellis for CASSANDRA-4572

+++ src/java/org/apache/cassandra/db/Directories.java
@@ -31,7 +31,6 @@ import org.slf4j.LoggerFactory;
 import org.apache.cassandra.config.*;
 import org.apache.cassandra.db.compaction.LeveledManifest;
 import org.apache.cassandra.io.util.FileUtils;
-import org.apache.cassandra.io.util.MmappedSegmentedFile;
 import org.apache.cassandra.io.sstable.*;
 import org.apache.cassandra.service.StorageService;
 import org.apache.cassandra.utils.CLibrary;
@@ -479,16 +478,24 @@ public class Directories
             if (!location.exists() || !location.isDirectory())
                 continue;
 
-            for (File ksDir : location.listFiles())
+            File[] ksDirs = location.listFiles();
+            if (ksDirs != null)
             {
-                if (!ksDir.isDirectory())
-                    continue;
+                for (File ksDir : ksDirs)
+                {
+                    if (!ksDir.isDirectory())
+                        continue;
 
-                for (File file : ksDir.listFiles())
-                    migrateFile(file, ksDir, null);
+                    File[] files = ksDir.listFiles();
+                    if (files != null)
+                    {
+                        for (File file : files)
+                            migrateFile(file, ksDir, null);
+                    }
 
-                migrateSnapshots(ksDir);
-                migrateBackups(ksDir);
+                    migrateSnapshots(ksDir);
+                    migrateBackups(ksDir);
+                }
             }
         }
     }
@@ -499,16 +506,23 @@ public class Directories
         if (!snapshotDir.exists())
             return;
 
-        for (File snapshot : snapshotDir.listFiles())
+        File[] snapshots = snapshotDir.listFiles();
+        if (snapshots != null)
         {
-            if (!snapshot.isDirectory())
-                continue;
+            for (File snapshot : snapshots)
+            {
+                if (!snapshot.isDirectory())
+                    continue;
 
-            for (File f : snapshot.listFiles())
-                migrateFile(f, ksDir, join(SNAPSHOT_SUBDIR, snapshot.getName()));
-
-            if (!snapshot.delete())
-                logger.info("Old snapsot directory {} not deleted by migraation as it is not empty", snapshot);
+                File[] files = snapshot.listFiles();
+                if (files != null)
+                {
+                    for (File f : files)
+                        migrateFile(f, ksDir, join(SNAPSHOT_SUBDIR, snapshot.getName()));
+                }
+                if (!snapshot.delete())
+                    logger.info("Old snapsot directory {} not deleted by migraation as it is not empty", snapshot);
+            }
         }
         if (!snapshotDir.delete())
             logger.info("Old directory {} not deleted by migration as it is not empty", snapshotDir);
@@ -520,9 +534,12 @@ public class Directories
         if (!backupDir.exists())
             return;
 
-        for (File f : backupDir.listFiles())
-            migrateFile(f, ksDir, BACKUPS_SUBDIR);
-
+        File[] files = backupDir.listFiles();
+        if (files != null)
+        {
+            for (File f : files)
+                migrateFile(f, ksDir, BACKUPS_SUBDIR);
+        }
         if (!backupDir.delete())
             logger.info("Old directory {} not deleted by migration as it is not empty", backupDir);
     }
